requirements.txt
Flask == 1.0.2
Pillow == 7.2.0
loguru == 0.5.3
matplotlib == 3.1.1
numpy == 1.19.5
pandas == 1.1.1
scikit-learn == 0.24.2
tensorflow == 2.4.1
tqdm == 4.64.0
transformers == 4.18.0
